wip: Move Model Selector to own file
